feat(refactor / ORM): fixed paths and set type for boolean
